+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Added
+
+- Widget: option to open collection on tap
+
 ## <a id="v1.7.1"></a>[v1.7.1] - 2022-10-09
 
 ### Added